Move test logger implementations to a submodule.
