+++ src/lib.rs
@@ -81,6 +81,11 @@ mod tokenize;
 pub mod data;
 pub mod tree;
 
+#[cfg(test)]
+pub use self::log::{
+    build_console_logger, build_json_logger, build_logger, build_null_logger,
+};
+
 pub use self::parse::{
     parse, ExtractedToken, ParseError, ParseErrorKind, ParseResult, Token,
 };
@@ -91,48 +96,4 @@ pub use self::tokenize::{tokenize, Tokenization};
 pub mod prelude {
     pub use super::tree::{Element, SyntaxTree};
     pub use super::{data, parse, preprocess, tokenize};
-}
-
-#[cfg(test)]
-#[inline]
-fn build_logger() -> slog::Logger {
-    build_null_logger()
-}
-
-#[cfg(test)]
-#[allow(dead_code)]
-fn build_null_logger() -> slog::Logger {
-    slog::Logger::root(slog::Discard, o!())
-}
-
-#[cfg(test)]
-#[allow(dead_code)]
-fn build_console_logger() -> slog::Logger {
-    use sloggers::terminal::TerminalLoggerBuilder;
-    use sloggers::types::Severity;
-    use sloggers::Build;
-
-    TerminalLoggerBuilder::new()
-        .level(Severity::Trace)
-        .build()
-        .expect("Unable to initialize logger")
-}
-
-#[cfg(test)]
-#[allow(dead_code)]
-fn build_json_logger() -> slog::Logger {
-    use slog::Drain;
-    use std::io;
-    use std::sync::Mutex;
-
-    // For writing to a file:
-    // + .add_default_keys()
-    // + .set_pretty(false)
-    let drain = slog_bunyan::with_name("ftml", io::stdout())
-        .set_newlines(true)
-        .set_pretty(true)
-        .set_flush(true)
-        .build();
-
-    slog::Logger::root(Mutex::new(drain).fuse(), o!("env" => "test"))
 }
